+++ click_extended/decorators/misc/__init__.py
@@ -1,5 +1,8 @@
 """Initialization file for the `click_extended.decorators.misc` module."""
 
+# pylint: disable=wrong-import-position
+
+from click_extended.decorators.misc.catch import catch
 from click_extended.decorators.misc.choice import choice
 from click_extended.decorators.misc.confirm_if import confirm_if
 from click_extended.decorators.misc.default import default
@@ -8,6 +11,7 @@ from click_extended.decorators.misc.experimental import experimental
 from click_extended.decorators.misc.now import now
 
 __all__ = [
+    "catch",
     "choice",
     "confirm_if",
     "default",